requirements.txt
fastapi~=0.73.0
python-dotenv~=0.19.2
SQLAlchemy~=1.4.31
alembic~=1.7.5
uvicorn~=0.17.0.post1
pydantic~=1.9.0
passlib~=1.7.4